Add `SolverSchedulePlugin` to encapsulate solver scheduling
